This add Boxed as a submodule to CycoD
Submodules are not my favorite thing but sometimes they just
seem to fit.  This may be one of those times.

We would likely want to add documentation about this once
we know the way we want people to play with this.

Or we don't use a submodule and just put this into CycoD
directly.