Update from Boxer submodule
This brings an important and useful change - it will now
auto-detect if you are on an Arm64 platform and default to using
that as the platform for the image.

This is important on Arm64 machines where the CPU emulation is
not 100% stable.  We have seen failures on Windows/WSL Linux
on Arm64 Windows systems where the linux/amd64 code would sometimes
fail due to some memory corruption.  The best guess is some issue
with multi-threaded memory access and the difference in memory
model between the Arm and Intel CPUs and the qemu emulation not
fully taking that into account.

Anyway, this should help those users to not have to manually override
the platform setting to use the native form.  The default should
now match Arm64 platforms.  (Just the default but that makes it
easier to use - all overrides and settings will continue to work
as before.)