/title refresh with continued conversation
Before: if your first command was `/title refresh` in a continued conversation, the file path would be null, and so the command wouldn't work.
Now: if your first command is `/title refresh` in a continued conversation, the original conversation's file path is used to read the conversation, then the new file path is only used to write, so the command works.