Refresh can now be used when title is locked
The original idea behind the title lock was to lock it from periodic, automatic AI changes. Over time though, it expanded to the idea of "lock from any and all changes". I can't really imagine a use case for this (why would the user want to lock the title from themselves?), so I'm reverting the idea back to the original: locking the title ONLY prevent automatic AI title changes. Right now of course, this only happens after the first interaction. But in the future, it could happen multiple times in a single conversation, ever X interactions, or whenever cycodev deems it necessary.