Added test chat client
This agent just returns test responses. These will be good for debugging and not wasting tokens. Also, cycodmd can use it to generate fake, static titles, which will be way easier to test, since we won't have to wait as long (or at all in a future implementation) for the response to come back.