Trim title generation context window
This will need some refinement, but it's better than having to wait upwards of 30 seconds for a title to be generated. Right now, it uses the first user and assistant messages, the last user and assistant messages, then whatever the 3rd and 4th to last messages in the conversation are. It might be too small of a context window, but it's still almost 10 seconds for a refresh, which is pretty large. Maybe I'll need to trim the actual messages themselves instead of just limiting the overall number of messages.