Refactored DisplayUserFunction
In the previous release (on master), DisplayUserFunctionCall was a private function, only called within ChatCommand.cs. The addition of multiple Command Handlers necessitated making the function public so that it was available to the new SlashCycoDmdCommandHandler class. This commit moves DisplayUserFunction and its sibling function DisplayGenericAssistantFunctionCall to ConsoleHelpers.cs, where other display functions are present already.