Refactor Persistent AI Provider Logic
Before, the logic to persist the current AI provider was loosely stuck in ApplyInstructions. Now there is a call to GetConfiguredAIProviders.

I also want to use this commit message to state that at this time, the Environment Variable solution will have to do. It works just fine for now, and since we're in the "garage" phase, I'm satisfied leaving it as is.