Catch FileLogger Failure during Title Generation
If FileLogger fails in subprocess, the warning message propagates back up to the title process, which uses that warning message as the title. This new logic catches this behavior and returns null. This should be improved upon in the future; improve subprocess logging to prevent contamination in the first place.